Finished basic states with command line UI
Basic state transfer is finished. Each state will transfer between the others
correctly. Only the base states have been finished. Admin needs to be added
to add more sodas to the machine as well as to track sales. The program currently has no
way to exit the loop as it is still in testing.