Fixed admin.java, added state transitions, and fixed stock tracking
Any time you enter 100 in a command prompt, you will go to the secret admin menu.
This allows restocking all of the sodas at once. It will also return the total amount of
everything that sold. You can leave that state to return to the NoQuarter. The project is
ready for the GUI to be added.